# docs/swagger.yaml
definitions:
  request.AppInstall:
    properties:
      name:
        type: string
      params:
        additionalProperties: true
        type: object
    type: object
  request.AppInstalledOperate:
    properties:
      action:
        type: string
      params:
        additionalProperties: true
        type: object
    type: object
  request.AppUnInstall:
    type: object
info:
  contact:
    email: xxyijixx@gmail.com
    name: xxyijixx
  description: Description of Doo Store API documentation
  license:
    name: AGPL-3.0
    url: https://opensource.org/license/agpl-v3
  title: Doo Store API Documentation
  version: "1.0"
paths:
  /apps:
    get:
      parameters:
      - default: 0
        description: page
        in: query
        name: page
        type: integer
      - default: 10
        description: pageSize
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: app page
      tags:
      - ""
  /apps/{key}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: key
        in: path
        name: key
        required: true
        type: string
      - description: RequestBody
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AppUnInstall'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: app uninstall
      tags:
      - ""
    post:
      consumes:
      - application/json
      parameters:
      - description: key
        in: path
        name: key
        required: true
        type: string
      - description: RequestBody
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AppInstall'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: app install
      tags:
      - ""
    put:
      consumes:
      - application/json
      parameters:
      - description: key
        in: path
        name: key
        required: true
        type: string
      - description: RequestBody
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AppInstalledOperate'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: app update
      tags:
      - ""
  /apps/{key}/detail:
    get:
      parameters:
      - description: key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: app detail
      tags:
      - ""
  /public/health:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: health
      tags:
      - public
securityDefinitions:
  BearerAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"